+++ Corona-Warn-App/src/main/java/de/rki/coronawarnapp/coronatest/worker/RatResultRetrievalWorker.kt
@@ -7,12 +7,12 @@ import dagger.assisted.Assisted
 import dagger.assisted.AssistedFactory
 import dagger.assisted.AssistedInject
 import de.rki.coronawarnapp.coronatest.CoronaTestRepository
-import de.rki.coronawarnapp.coronatest.execution.TestResultScheduler
-import de.rki.coronawarnapp.coronatest.execution.TestResultScheduler.RatPollingMode.DISABLED
-import de.rki.coronawarnapp.coronatest.execution.TestResultScheduler.RatPollingMode.PHASE1
-import de.rki.coronawarnapp.coronatest.execution.TestResultScheduler.RatPollingMode.PHASE2
 import de.rki.coronawarnapp.coronatest.latestRAT
 import de.rki.coronawarnapp.coronatest.type.CoronaTest
+import de.rki.coronawarnapp.coronatest.worker.execution.RatResultScheduler
+import de.rki.coronawarnapp.coronatest.worker.execution.RatResultScheduler.RatPollingMode.DISABLED
+import de.rki.coronawarnapp.coronatest.worker.execution.RatResultScheduler.RatPollingMode.PHASE1
+import de.rki.coronawarnapp.coronatest.worker.execution.RatResultScheduler.RatPollingMode.PHASE2
 import de.rki.coronawarnapp.util.TimeStamper
 import de.rki.coronawarnapp.util.worker.InjectedWorkerFactory
 import de.rki.coronawarnapp.worker.BackgroundConstants
@@ -28,7 +28,7 @@ class RatResultRetrievalWorker @AssistedInject constructor(
     @Assisted workerParams: WorkerParameters,
     private val coronaTestRepository: CoronaTestRepository,
     private val timeStamper: TimeStamper,
-    private val testResultScheduler: TestResultScheduler,
+    private val ratResultScheduler: RatResultScheduler,
 ) : CoroutineWorker(context, workerParams) {
 
     override suspend fun doWork(): Result {
@@ -37,7 +37,7 @@ class RatResultRetrievalWorker @AssistedInject constructor(
         if (runAttemptCount > BackgroundConstants.WORKER_RETRY_COUNT_THRESHOLD) {
             Timber.tag(TAG).d("$id doWork() failed after $runAttemptCount attempts. Rescheduling")
 
-            testResultScheduler.setRatResultPeriodicPollingMode(mode = testResultScheduler.ratResultPeriodicPollingMode)
+            ratResultScheduler.setRatResultPeriodicPollingMode(mode = ratResultScheduler.ratResultPeriodicPollingMode)
             Timber.tag(TAG).d("$id Rescheduled background worker")
 
             return Result.failure()
@@ -53,7 +53,7 @@ class RatResultRetrievalWorker @AssistedInject constructor(
             val nowUTC = timeStamper.nowUTC
             val days = Duration(rat.registeredAt, nowUTC).standardDays
             val minutes = Duration(rat.registeredAt, nowUTC).standardMinutes
-            val isPhase1 = testResultScheduler.ratResultPeriodicPollingMode == PHASE1
+            val isPhase1 = ratResultScheduler.ratResultPeriodicPollingMode == PHASE1
             Timber.tag(TAG).d("Calculated days: %d", days)
             when {
                 rat.isResultAvailableNotificationSent -> {
@@ -70,7 +70,7 @@ class RatResultRetrievalWorker @AssistedInject constructor(
                 }
                 isPhase1 && minutes >= RAT_POLLING_END_OF_PHASE1_MINUTES -> {
                     Timber.tag(TAG).d("$id $minutes minutes - time for a phase 2!")
-                    testResultScheduler.setRatResultPeriodicPollingMode(mode = PHASE2)
+                    ratResultScheduler.setRatResultPeriodicPollingMode(mode = PHASE2)
                 }
                 else -> {
                     coronaTestRepository.refresh(CoronaTest.Type.RAPID_ANTIGEN)
@@ -81,7 +81,7 @@ class RatResultRetrievalWorker @AssistedInject constructor(
     }
 
     private fun disablePolling() {
-        testResultScheduler.setRatResultPeriodicPollingMode(mode = DISABLED)
+        ratResultScheduler.setRatResultPeriodicPollingMode(mode = DISABLED)
         Timber.tag(TAG).d("$id: Background worker stopped")
     }
