+++ Corona-Warn-App/src/main/java/de/rki/coronawarnapp/coronatest/worker/execution/PCRTestResultScheduler.kt
@@ -1,17 +1,13 @@
-package de.rki.coronawarnapp.coronatest.execution
+package de.rki.coronawarnapp.coronatest.worker.execution
 
 import androidx.annotation.VisibleForTesting
 import androidx.work.BackoffPolicy
 import androidx.work.ExistingPeriodicWorkPolicy
-import androidx.work.PeriodicWorkRequest
 import androidx.work.PeriodicWorkRequestBuilder
 import androidx.work.WorkInfo
 import androidx.work.WorkManager
 import dagger.Reusable
-import de.rki.coronawarnapp.coronatest.execution.TestResultScheduler.RatPollingMode.DISABLED
-import de.rki.coronawarnapp.coronatest.execution.TestResultScheduler.RatPollingMode.PHASE1
 import de.rki.coronawarnapp.coronatest.worker.PCRTestResultRetrievalWorker
-import de.rki.coronawarnapp.coronatest.worker.RatResultRetrievalWorker
 import de.rki.coronawarnapp.util.coroutine.await
 import de.rki.coronawarnapp.worker.BackgroundConstants
 import de.rki.coronawarnapp.worker.BackgroundWorkHelper
@@ -21,13 +17,9 @@ import java.util.concurrent.TimeUnit
 import javax.inject.Inject
 
 @Reusable
-class TestResultScheduler @Inject constructor(
+class PCRTestResultScheduler @Inject constructor(
     private val workManager: WorkManager,
 ) {
-
-    private var ratWorkerMode = DISABLED
-    val ratResultPeriodicPollingMode
-        get() = ratWorkerMode
 
     private suspend fun isPcrScheduled() =
         workManager.getWorkInfosForUniqueWork(PCR_TESTRESULT_WORKER_UNIQUEUNAME)
@@ -57,29 +49,6 @@ class TestResultScheduler @Inject constructor(
         }
     }
 
-    enum class RatPollingMode {
-        DISABLED,
-        PHASE1,
-        PHASE2
-    }
-
-    fun setRatResultPeriodicPollingMode(mode: RatPollingMode) {
-        ratWorkerMode = mode
-        if (mode == DISABLED) {
-            Timber.tag(TAG).d("cancelWorker()")
-            workManager.cancelUniqueWork(RAT_RESULT_WORKER_UNIQUEUNAME)
-        } else {
-            // no check for already running workers!
-            // worker must be replaced by next phase instance
-            Timber.tag(TAG).i("Queueing rat result worker (RAT_RESULT_PERIODIC_WORKER)")
-            workManager.enqueueUniquePeriodicWork(
-                RAT_RESULT_WORKER_UNIQUEUNAME,
-                ExistingPeriodicWorkPolicy.REPLACE,
-                buildRatResultRetrievalPeriodicWork(mode)
-            )
-        }
-    }
-
     private fun buildPcrTestResultRetrievalPeriodicWork() =
         PeriodicWorkRequestBuilder<PCRTestResultRetrievalWorker>(
             getPcrTestResultRetrievalPeriodicWorkTimeInterval(),
@@ -97,29 +66,6 @@ class TestResultScheduler @Inject constructor(
             )
             .build()
 
-    private fun buildRatResultRetrievalPeriodicWork(pollingMode: RatPollingMode): PeriodicWorkRequest {
-        val repeatInterval = if (pollingMode == PHASE1) {
-            ratResultRetrievalPeriodicWorkPhase1IntervalInMinutes
-        } else {
-            ratResultRetrievalPeriodicWorkPhase2IntervalInMinutes
-        }
-        return PeriodicWorkRequestBuilder<RatResultRetrievalWorker>(
-            repeatInterval,
-            TimeUnit.MINUTES
-        )
-            .addTag(RAT_RESULT_WORKER_TAG)
-            .setConstraints(BackgroundWorkHelper.getConstraintsForDiagnosisKeyOneTimeBackgroundWork())
-            .setInitialDelay(
-                DIAGNOSIS_TEST_RESULT_PERIODIC_INITIAL_DELAY,
-                TimeUnit.SECONDS
-            ).setBackoffCriteria(
-                BackoffPolicy.LINEAR,
-                BackgroundConstants.KIND_DELAY,
-                TimeUnit.MINUTES
-            )
-            .build()
-    }
-
     companion object {
         /**
          * Kind initial delay in minutes for periodic work for accessibility reason
@@ -128,11 +74,9 @@ class TestResultScheduler @Inject constructor(
          */
         private const val DIAGNOSIS_TEST_RESULT_PERIODIC_INITIAL_DELAY = 10L
         private const val PCR_TESTRESULT_WORKER_TAG = "DIAGNOSIS_TEST_RESULT_PERIODIC_WORKER"
-        private const val RAT_RESULT_WORKER_TAG = "RAT_RESULT_PERIODIC_WORKER"
         private const val PCR_TESTRESULT_WORKER_UNIQUEUNAME = "DiagnosisTestResultBackgroundPeriodicWork"
-        private const val RAT_RESULT_WORKER_UNIQUEUNAME = "RatResultRetrievalWorker"
 
-        private const val TAG = "TestResultScheduler"
+        private const val TAG = "PCRTestResultScheduler"
 
         /**
          * Calculate the time for pcr diagnosis key retrieval periodic work
@@ -147,9 +91,5 @@ class TestResultScheduler @Inject constructor(
                 BackgroundConstants.MINUTES_IN_DAY /
                     BackgroundConstants.DIAGNOSIS_TEST_RESULT_RETRIEVAL_TRIES_PER_DAY
                 ).toLong()
-
-        private const val ratResultRetrievalPeriodicWorkPhase1IntervalInMinutes = 15L
-
-        private const val ratResultRetrievalPeriodicWorkPhase2IntervalInMinutes = 90L
     }
 }
