+++ Corona-Warn-App/src/main/java/de/rki/coronawarnapp/worker/BackgroundWorkScheduler.kt
@@ -1,7 +1,7 @@
 package de.rki.coronawarnapp.worker
 
 import de.rki.coronawarnapp.coronatest.CoronaTestRepository
-import de.rki.coronawarnapp.coronatest.execution.TestResultScheduler
+import de.rki.coronawarnapp.coronatest.worker.execution.PCRTestResultScheduler
 import de.rki.coronawarnapp.deniability.NoiseScheduler
 import de.rki.coronawarnapp.risk.execution.RiskWorkScheduler
 import kotlinx.coroutines.flow.first
@@ -21,7 +21,7 @@ import javax.inject.Singleton
 class BackgroundWorkScheduler @Inject constructor(
     private val riskWorkScheduler: RiskWorkScheduler,
     private val coronaTestRepository: CoronaTestRepository,
-    private val testResultScheduler: TestResultScheduler,
+    private val testResultScheduler: PCRTestResultScheduler,
     private val noiseScheduler: NoiseScheduler,
 ) {
